requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg==3.2.1
psycopg-binary==3.2.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1

django-debug-toolbar~=4.4.6
Faker~=28.0.0
python-dotenv == 1.0.1
gunicorn == 21.2.0